property composition
Different kinds of events (here, sync and io) are modelled in separate classes and
composed in an vertex_event_property class.

This commit also wraps the sync and io event property in a variant.
This enables us to have just one member being either an io or an sync
event type. The enum `vertex_kind` represents what kind of event we get
when accessing the `property` member.