[WIP] add generic graph builder
This commit provides a `make_graph` function, which gets a Builder class
as template type parameter. The Builder class must be a callable object.
Furthermore, the `make_graph` function gets the arguments as parameter
pack to call the builder. It afterwards returns the resulting graph from
the builder.

The graph which, is hold by the `simple_graph_builder` class is now a
std::unique_ptr to easily transfer the ownership of the graph to the
caller of `make_graph` after the graph is created.