[WIP] Refactor visitor algorithm
As a container we use now a `std::map` which has a process id as a key and a
vector of `CIO_Sets` for each process as `mapped_type`.

`CIO_Set` does not have a state map where the state of each process is
stored. Instead just a `Set_State` is stored.

The `CIO_Visitor` just gather the I/O Sets per Process. The final sets
will be build in a post-processing step after the dfs. This final step
is not content of this commit.

The `start_event` and `end_event` of a set is now the root event of the
synchronization no longer the synchronization event for this process
itself. This is important for the furthermore algorithm.