Set-per-Process algorithm
* If the DFS visitation starts with an process which has
root-synchronization-events, the algorithms starts backtracking.
With the former, naive implementation start and end events of the other
processes break. This commit adds some additional checks on the edge
examination to tackle these problems.

* First for the synthetic end event we have just one visitation of the
actual vertex. Afterwards the vertex is marked visited and all other
processes are just visiting the edge.

* Second if we backtrack sync-events the end-event of a synchronization
is visited by the process doing the backtracking before the start event.
For that we also have to close the set if we are on the edge to the end
event.

* Third if we come from the synthetic root event we have to open a new
set. Since synthetic event have no pid we do this when visiting the edge
and take the pid from the target vertex.
If the target is a sync event, we close the set immediately.

* Tested on three different Graphs.