Merge concurrent I/O Sets together
This commit provides the Algorithmen to merge per-process-sets together
to Concurrent I/O Sets. This includes:

* Filtering of empty sets:
The remove_empty_sets function deletes all empty Per-Process Sets.
This can happen if two consecutive Sync-Events appear in the Graph.

* Getting all syncronization events in the graph:
It seems to be simpler getting all sync-events directly from the graph
instead of filter them out of the per-process-set-map.

* Sort sets by Descriptor:
Through backtracking it is possible that the processes hold its sets not
in depth-first-order. Since we rely on depth-first-order in the
following here we sort by vertex_descriptor id.

* Collecting all root synchronization events:
For each sync-event occuring in the graph we determine the root.
Then sort them by descriptor ID to delete the duplicates afterwards.

* Sort events chronological:
For the merging of the concurrent I/O sets we need a list of
chrnological sorted end-events of the per-process sets.

* Merge Sets:
Here we take the first set of each process. Copy all event in these set
into a new set. The choosing the end-event of the per-process-set which
appears chronological first (has the lowest timestamp). And delete these
per-process-sets from the set-map.
This is repeated until the set-map doesn't contains any sets anymore.