Provide synthetic event type as type
Until now the `synthetic` `vertex_kind` was just a kind with no
corresponding type other than `sync_event` or `io_event` with this
commit this changes. The `synthetic_event_property` class holds a name
and a timestamp, this enables to distinguish between root and end by
providing different names. Also a synthetic timestamp to guarantee a
correct temporal ordering of possible start- and end-events in a set.
Since even synthetic events provide timestamps the api for accessing the
timestamp from an `otf2_trace_event` has changed. Instead of a optional
we can return the `time_point` directly.