handle synthetic events properly
* provide `get_events_by_kind`
Since synthetic events can be also end-events of a set we have to handle
them properly in our implementation.
This commit provides a `get_events_by_kind` function for gathering
different types of events in any combination.

* Change in `root_of_sync`:
The root of a synthetic event is always the event itself.

* Provide a `std::vector` overload of the `remove_empty_sets` function.
Since we have to remove the empty sets after merging instead before.