add `sync_scope` to classify sync-events
* Adding a `sync_scope` enum class to classify sync-events.
We distinguish into two different classes of sync-events,
'Local' and 'Global' where local means that the synchronization appears
not over all processes in the program, instead just on a subgroup of
them. Global respectively means all Processes, executing the programm
are involved in the synchronization.

* Further the corresponding ostream overload is provided for debugging
and printing.

* The `classify_sync` functions comes with two different overloads.
One, acting on the graph and a vertex descriptor.
The seconde takes the graph and a `sync_event_property` type.