update merge_sets algorithm
* First with this commit new type-aliases are introduced.
* Remove namespace-prefix where it is not necessary.

* Introduce `map_view` type it represents a view to a set_map.
It consits of the proc_id (uint64_t) as key type and a pair of iterators.
The iterators pointing to the begin and the end of the set-vector of the process
referenced by key. This is quite danagerous, if the underlying
`set_map_t` object which holds the vectors of sets the iterators
pointing to, goes out of scope, then iterators are invalidated!
The reason to use this sort of view-like type is that I assume its
cheapter to copy.
The recursive algorithm needs to copy the `map_view` after updating them.

* Provide `make_mapview` generator function to get a `map_view_t` type
from a `set_map_t` type.

* Add `update_view` function wich takes a `map_view_t` type per value
incrementing its iterators by a given `process_group_t` object.
And returning the updated `map_view_t` object per value.
This is used to pass an updated `map_view_t` when calling `process_sets`
recursively.

* Add a `pg_group` function which takes a `Graph` reference and a
`VertexDescriptor` and returns a corresponding `process_group_t` object,
which consists of the process Ids of all processes corresponding to the
event of the Vertex
There is currently no implementation for the case that the referenced
Vertex is an `io_event_property`.

* The `do_merge` function takes a const-reference to a `map_view_t`
object and a non-const-referenced for the `merged_sets` vector.
It creates a temporary `set_t` object and copy all events of the sets
referenced by the begin iterator of the current process, while looping
over the map_view. Afterwards the set is closed and pushed back into the
`merged_sets` vector. TODO: The selection of the end_evt which will be
set for the new set is not correct.

* The `independent_end_evts` function is work in progress.
We get a copy of the end events vector, sort them by ascending
`VertexDescriptor`. Then all duplicates will be removed. If the
resulting size of the end event vector is 1, return an empty vector,
since there is no dependency between a single sync-event.
Otherwise we loop over each permutation of end events picking two
end events get there process groups and build the intersection of these
process groups. If the result of the intersection is empty we push back
a `pair` object of the end events into the result vector.

* The `process_sets` function is the recursive core of the
`detail::merge_sets_impl` algorithm. It takes a `map_view_t` object per
value, the `merged_sets` result vector as non-const-reference and a
const-reference to the sorted-root-sync events.
In the case that all begin iterators in the `map_view_t` object are
equal with its end iterators, return.
Otherwise merge the "current" view into a new set and get a vector
of all end events of these sets. Then getting a vector of the independent end
events. If the vector is empty, take the first one. Afterwards update
the map_view respectively to the process group of the end event and go
into recursion.
Else iterate over the the independent end event pairs and updating the
map view for each of them and call `process_sets` recursive.

* The `merge_sets_impl` lives in the detail namespace.
Here we basically create the map view of a `set_map_t`
than calling `process_sets`.

* The old algorithm is also still accessible through the detail
namespace. as `merge_sets_impl_old`.

* The public `merge_sets` function just call `detail::merge_sets_impl`.