WIP - set origin end-event, sort view, ...
* Update `CIO_Set`, save the origin sync event of the corresponding
process of the set when create the per-process-sets. The `origin` method returns
this event of type `value_type`. An overwirte for the `set_end_event`
method which, takes the `root_end_event` and the `origin` and sets them
properly.
Default initialize `origin` with `value_type` max in numeric limits.
Note: All this `CIO_Set`-stuff works only for vertex descriptors of
integral type.

* Set origin end-event in `CIO_Visitor` through the new overload.

* Add a sort function for a `set_map_t` type, which takes the timestamp
of the `origin` end-event of the set. This is needed to get a `view` of
the sync events like they occur in the graph.
WORK IN PROGRESS!
TODO: It might be better to use an adjacency relation of the graph.
Nevertheless, this should also work, since events of the same process
can not overhaul themselve.

* The utility function `generate_unique_pairs` is used to get all
unique 2-pairs of combination of a vector of end-events, which must
be of an integral type.

* Rename `independent_end_evts` to `independent_sync_pairs` which is
more descriptive for what it is. Instead of looping over all
permutations of the end-event vector we use the `generate_unique_pairs`
function to get all pairs of end-events  If the intersection of the
process groups of these end-event pairs are empty the two events are
indendent of each other. Since they affecting other processes then the
other one.

* Update `process_sets`, using `independent_sync_pairs` and check if we
really can update the end-event. This is WORK IN PROGRESS!