find end events to update
This is a work-in-progress commit and might be rebased later.

* Add `find_end_events_to_update` function. This function should return
one or more end-events which are possible to update from the view.
As arguments it takes a Graph non-const-reference and a copy of the
end-events vector.
Finding the end-events to update works as follows:
1. Capture the end-events vector by copy in a `check_update_func`
   and a `check_update_func_single` closure. This is work in progress,
   not sure if we need both of them.
   The `check_update_func` closure gets a vector of possible end-events
   and returns the first end-event which can be updated from the
   end-event vector.
   The `check_update_func_single` closure gets just a single end-event
   and returns a bool indicating whether the given end-event can be
   updated or not.
2. After we have the end-event vector captured in the closures as it is.
Then we sort them and make them unique to reduce the size.
If just one element we can assume that we are on a global
synchronization event so we can safely return the one-element-vector for
update. Otherwise no global synchronization event can be updated even if
it the vector contains one. Therfore we delete all erase all
global-synchronization-events from the vector. If the remaining
end-events vector size is just one now? Then, return the vector.
With just one end-event we cannot depent on any other so just return it.
In the case that there are still more than two end-events, the regular
case. Call `generate_unique_pairs` of the end-event vector and check
each pair if its indpendent from each other. A pair of end-events is
independet when none of the processes corresponding in the
synchronization-event overlap. In the case a pair is independent we
insert each event of the pair in the `independent_syncs` set. And in the
case that they are dependet we insert them in the `dependent_syncs` set.
TODO: At the moment we also store the overlapping processes. This seems
not useless.
3. `dependent_syncs` is not empty. In this case we have dependet
   synchronization-events in the end-event vector. So we pick the first
   we can update and update them.
4. `independent_syncs` is not empty. In this case we have independet
   synchronization-events in the end-event vector. Here, we return a
   vector of each independet synchronization-event which can be updated.

* In the `process_sets` function we had nothing else to do than to loop over
the end-event which are ready for update, and call recursive with
an updated map.

* Get rid of `sorted_syncs` vector during the whole algorithm. To rely
on the chronologically sorted root-sync-events does not work because we
cannot rely on the timestamp of the synchronization-root-events.

* The `do_merge` function takes a Graph non-const-reference and instead
of setting the end-event of the new set by searching for the earliest
occurence of one of the end-events of the current view, we call the new
`find_end_events_to_update` function and receive a vector of possible
end-events. Here we should take the one we later drop from the view by
update. At the moment we simply use the `back` of the vector.

* Fix order of `can_update_end_event` function.
Here the problem was, we have an overload with an second template
parameter for the first argument to a function. So we have to place the
definition of the overload which takes just one template-parameter and
as a first argument a `process_group_t` object first. Otherwise the
`Graph` template-parameter will also deduce to `process_group_t` and
this is not what we want.

* Replace the `independent_sync_pairs` function entirely with
`find_end_events_to_update` function.