fix test case for concurrent io sets
* Since we compare a vector of sets of ints with an vector of cio_sets
of `vertex_descriptor` type, which must be integral. But we can't rely
on any sorting within that, we have to satisfy that all cio_sets in our
cio-set vector has any set in the expected-set vector where, all of the
element within that cio-sets occur anywhere in the expected-set
vector.