[WIP] distinguish otf2 io-event kinds
* To be able to build meaningful analysis of the cio-sets. It is
necessary to provide a better way to distinguish the different
I/O-events within a CIO-set. For that reason, this commit introduces
a new enum class wich describe the kind of I/O-event.
At the moment we distinguish the following kinds of I/O-events.
*  create,              // create or open a new handle
*  dup,                 // duplicate a handle
*  seek,                // seek
*  read,                // read operation
*  write,               // write operation
*  flush,               // flush
*  delete_or_close,     // delete or close handle
*  none                 // none
The kind will be set at construction type, inside the graph builder.