[WIP] Delete old complicated functionality
This commit removes the following functions:
    1. `get_events_by_kind`
    2. `get_io_events_by_kind`
    3. `make_local_pgmap`
    4. `collect_root_sync_events`
The first two `get_*` filter properties of the graph which is not really
useful. Since we have the sets as starting point.
The `make_local_pgmap` function was completetly unused execpt in test
cases. Possibly reimplement them locally in the testfile.
The `collect_root_sync_events` function is waste.