[WIP] Provide duration of events
We measure the duration of events with
leave_event.timestamp() - enter_event.timestamp().
Therefore a stack is maintained storing the enter timestamp, a duration
and an optional vertex descriptor if the duration of the current
operation should be stored inside the vertex.

The duration is stored inside the `otf2_trace_event` vertex property.

Additional we try to measure the total time of the application by adding
the duration of the single operations together inside the leave
callback.

Furthermore by distinguish the `role_type` of the event we measure the
file_io time and file_io_metadata time.

This functionallity is not tested.