Fix region stack
The region stack that was used before was wrong.
The location_queue class used a deque under the hood and therfore
if the depth of the stack is greater than one region names where
assigned wrong.

This commit adds a `location_stack` class similar to the
`location_queue` but with a `std::stack` as underlying container.
Even the interface of `std::stack` is adaped by this class.