[WIP] eliminate recursion
* Transform recursion into iteration in `process_sets`.
This change is made to avoid stackoverflows with larger trace files.
Instead of updating the view and make a recursive call for each possible update.
The possible views are stored in a vector which will be processed iteratively.

* The `process_sets` function returns the vector of resulting
CIO-Sets instead of getting an out reference parameter.

* The `do_merge` function returns the newly created CIO-Set and the
end-events vector for this step.

* The `make_current_view` function returns a vector containing the
iterators to the current CIO-Set of the view of each process.
This makes iteration simpler.

* Disable the overlapping_procs vector in `can_update_end_event` since
its just useful for debugging.