[WIP] add hash to dataframe
In order to mark events of interest with marker in the trace file.
This commit adds a hash value to the dataframe.
The hash value consists out of the (pid, kind, timestamp) of an event.
In the first step, the hash is just added to the `Intervals` of the
overlapping access analysis.